backend/src/users/users.service.ts
import { Injectable, Inject, forwardRef } from '@nestjs/common';
import { InjectRepository } from '@nestjs/typeorm';
import { Repository } from 'typeorm';
import { User } from './user.entity';
import { AuthService } from '../auth/auth.service';

@Injectable()
export class UsersService {
  constructor(
    @InjectRepository(User)
    private usersRepository: Repository<User>,
    @Inject(forwardRef(() => AuthService))
    private authService: AuthService,
  ) {}

  async findOrCreate(username: string): Promise<User> {
    let user = await this.usersRepository.findOne({ where: { username } });
    if (!user) {
      user = this.usersRepository.create({ username });
      await this.usersRepository.save(user);
    }
    return user;
  }

  async findById(id: number): Promise<User> {
    return this.usersRepository.findOne({ where: { id } });
  }
}
